Lots of rework
Only the provider group remains

+++ src/commands/images/emojis.js
@@ -15,6 +15,18 @@
  *   along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+/**
+ * @file images EmojisCommand - Sends a png image along with a message to fake sending global emojis  
+ * **Aliases**: `emoji`, `emo`, `sendemo`, `emosend`, `sendemoji`
+ * @module
+ * @category images
+ * @name emojis
+ * @example sendemoji thonk
+ * @param {StringResolvable} EmojiName Name of the emoji to send
+ * @param {StringResolvable} [Content] Optional content to send along with the emoji
+ * @returns {Message} Emoji image and optional content
+ */
+
 const dym = require('didyoumean2'),
   fs = require('fs'),
   path = require('path'),
@@ -22,14 +34,14 @@ const dym = require('didyoumean2'),
   {oneLine} = require('common-tags'),
   {deleteCommandMessages} = require(path.join(__dirname, '../../util.js'));
 
-module.exports = class sendEmojiCommand extends Command {
+module.exports = class EmojisCommand extends Command {
   constructor (client) {
     super(client, {
       name: 'emojis',
       memberName: 'emojis',
       group: 'images',
       aliases: ['emoji', 'emo', 'sendemo', 'emosend', 'sendemoji'],
-      description: 'Send an emoji',
+      description: 'Sends a png image along with a message to fake sending global emojis',
       format: 'EmojiName [MessageToSendWithEmoji]',
       examples: ['sendemoji thonk'],
       guildOnly: false,
@@ -50,17 +62,16 @@ module.exports = class sendEmojiCommand extends Command {
   }
 
   run (msg, {emoji, message}) {
-    try {
+    if (fs.existsSync(path.join(__dirname, `../../data/images/emojis/${emoji}.png`))) {
       deleteCommandMessages(msg, this.client);
 
       return msg.say(message, {files: [path.join(__dirname, `../../data/images/emojis/${emoji}.png`)]});
-    } catch (err) {
-      const matchList = fs.readdirSync(path.join).map(v => v.slice(0, 4)),
-        maybe = dym(emoji, matchList, {deburr: true});
-
-      return msg.reply(oneLine`that emoji does not exist! ${maybe 
-        ? oneLine`Did you mean \`${maybe}\`?` 
-        : 'You can add it to the folder then try again'}`);
     }
+    const matchList = fs.readdirSync(path.join(__dirname, '../../data/images/emojis/')).map(v => v.slice(0, 4)),
+      maybe = dym(emoji, matchList, {deburr: true});
+
+    return msg.reply(oneLine`that emoji does not exist! ${maybe 
+      ? oneLine`Did you mean \`${maybe}\`?` 
+      : 'You can add it to the folder then try again'}`);
   }
 };
